rename collider -> colliders
