+++ cocos/physics/framework/deprecated.ts
@@ -4,11 +4,11 @@
 
 import { PhysicsSystem } from "./physics-system";
 import { replaceProperty, removeProperty } from "../../core";
-import { BoxColliderComponent } from "./components/collider/box-collider-component";
-import { SphereColliderComponent } from "./components/collider/sphere-collider-component";
-import { CapsuleColliderComponent } from "./components/collider/capsule-collider-component";
+import { BoxColliderComponent } from "./components/colliders/box-collider-component";
+import { SphereColliderComponent } from "./components/colliders/sphere-collider-component";
+import { CapsuleColliderComponent } from "./components/colliders/capsule-collider-component";
 import { RigidBodyComponent } from "./components/rigid-body-component";
-import { ColliderComponent } from "./components/collider/collider-component";
+import { ColliderComponent } from "./components/colliders/collider-component";
 
 replaceProperty(PhysicsSystem, 'PhysicsSystem', [
     {
@@ -17,7 +17,7 @@ replaceProperty(PhysicsSystem, 'PhysicsSystem', [
     }
 ]);
 
-replaceProperty(PhysicsSystem.prototype, 'PhysicsSystem', [
+replaceProperty(PhysicsSystem.prototype, 'PhysicsSystem.prototype', [
     {
         "name": "deltaTime",
         "newName": "fixedTimeStep"
@@ -28,7 +28,7 @@ replaceProperty(PhysicsSystem.prototype, 'PhysicsSystem', [
     }
 ]);
 
-removeProperty(PhysicsSystem.prototype, 'PhysicsSystem', [
+removeProperty(PhysicsSystem.prototype, 'PhysicsSystem.prototype', [
     {
         "name": "useFixedTime"
     }
