+++ cocos/physics/framework/index.ts
@@ -5,18 +5,18 @@
 import { PhysicsSystem } from './physics-system';
 import { PhysicMaterial } from './assets/physic-material';
 import { PhysicsRayResult } from './physics-ray-result';
-import { BoxColliderComponent } from './components/collider/box-collider-component';
-import { ColliderComponent } from './components/collider/collider-component';
-import { SphereColliderComponent } from './components/collider/sphere-collider-component';
-import { CapsuleColliderComponent } from './components/collider/capsule-collider-component';
-import { CylinderColliderComponent } from './components/collider/cylinder-collider-component';
-import { ConeColliderComponent } from './components/collider/cone-collider-component';
-import { MeshColliderComponent } from './components/collider/mesh-collider-component';
+import { BoxColliderComponent } from './components/colliders/box-collider-component';
+import { ColliderComponent } from './components/colliders/collider-component';
+import { SphereColliderComponent } from './components/colliders/sphere-collider-component';
+import { CapsuleColliderComponent } from './components/colliders/capsule-collider-component';
+import { CylinderColliderComponent } from './components/colliders/cylinder-collider-component';
+import { ConeColliderComponent } from './components/colliders/cone-collider-component';
+import { MeshColliderComponent } from './components/colliders/mesh-collider-component';
 import { RigidBodyComponent } from './components/rigid-body-component';
 import { ConstantForce } from './components/constant-force';
-import { TerrainColliderComponent } from './components/collider/terrain-collider-component';
-import { SimplexColliderComponent } from './components/collider/simplex-collider-component';
-import { PlaneColliderComponent } from './components/collider/plane-collider-component';
+import { TerrainColliderComponent } from './components/colliders/terrain-collider-component';
+import { SimplexColliderComponent } from './components/colliders/simplex-collider-component';
+import { PlaneColliderComponent } from './components/colliders/plane-collider-component';
 
 export {
     PhysicsSystem,
